+++ src/hooks/quries/course.ts
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query'
+import { getCourseInformationApi } from '@/api/courseInformation'
+
+export const useGetCourses = () => {
+  return useQuery({
+    queryKey: ['courses'],
+    queryFn: getCourseInformationApi,
+  })
+}
